refactor: Modularize view.ts file
