+++ src/blocks/upload-song/ts/configs.ts
@@ -1,0 +1,7 @@
+// Configuration values for the plugin
+const configs = {
+	apiPath: '/wp-json/songwriter-tools/v1', // Adjust this to match your API endpoint
+	storeNamespace: 'upload-block',
+};
+
+export default configs;
